python--pingdom-api-integration/requirements.txt
requests
PyYaml
